all: fixed make test issues
