+++ client/internal/assets/test.json
@@ -1,7 +1,7 @@
 {
   "config": {
     "chainID": 3680923,
-    "tendermint": {
+    "konsensus": {
       "rewarded": true
     }
   },
