+++ client/knode/genesis/testfiles/testnet_default.json
@@ -1,7 +1,7 @@
 {
   "config": {
     "chainID": 1000,
-    "tendermint": {
+    "konsensus": {
       "rewarded": true
     }
   },
